gitconfig: set crlf = input
If you’re on a Linux or macOS system that uses LF line endings, then you
don’t want Git to automatically convert them when you check out files;
however, if a file with CRLF endings accidentally gets introduced, then
you may want Git to fix it. You can tell Git to convert CRLF to LF on
commit but not the other way around by setting core.autocrlf to input:

$ git config --global core.autocrlf input

This setup should leave you with CRLF endings in Windows checkouts, but
LF endings on macOS and Linux systems and in the repository.